add the following features: -add picasso feature to load image from server -add refresh option to get temp of last city -covert temperature degree from kelvin to celsius -remove unused drawables

+++ app/src/main/java/com/elfalt/weatherforecast/models/WeatherResponse.kt
@@ -2,4 +2,4 @@ package com.elfalt.weatherforecast.models
 
 data class WeatherResponse(val name: String, val weather: List<WeatherInfo>, val main : Weather)
 
-data class Weather(val temp : Double)
+data class Weather(val temp : Float)
